+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #  Change Log
 
 ## `master`
+* Fix `TileJSON` init parameters
 
 ## 0.1.1
 * Add more documentation